refactor: modify project structure
