+++ internal/auth/keycloak.go
@@ -1,4 +1,4 @@
-package keycloak
+package auth
 
 import (
 	"errors"
