--- internal/clients/impartus.go
@@ -1,5 +1,0 @@
-package clients
-
-import pkg "github.com/crux-bphc/lex/pkg/impartus"
-
-var ImpartusClient = pkg.ImpartusClient{BaseUrl: "https://bitshyd.impartus.com/api"}
