+++ internal/impartus/client.go
@@ -1,0 +1,5 @@
+package impartus
+
+import pkg "github.com/crux-bphc/lex/pkg/impartus"
+
+var Client = pkg.ImpartusClient{BaseUrl: "https://bitshyd.impartus.com/api"}
