+++ internal/routes/user.go
@@ -1,20 +1,20 @@
 package routes
 
 import (
-	"github.com/crux-bphc/lex/internal/clients/keycloak"
+	"github.com/crux-bphc/lex/internal/auth"
 	"github.com/gin-gonic/gin"
 )
 
 func RegisterUserRoutes(router *gin.Engine) {
 	r := router.Group("/user")
 
-	r.Use(keycloak.Middleware())
+	r.Use(auth.Middleware())
 
 	// TODO: rudimentary endpoint for testing
 	r.GET("/info", func(ctx *gin.Context) {
 
 		ctx.JSON(200, gin.H{
-			"claims": keycloak.GetClaims(ctx),
+			"claims": auth.GetClaims(ctx),
 		})
 	})
 }
